% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrDEG.R
\name{corrDEG}
\alias{corrDEG}
\title{Finds the correlation pre/post batch correction of the DEG by cluster.}
\usage{
corrDEG(DEG_df)
}
\arguments{
\item{DEG_df}{data frame with DEG information for each batch pre and
post correction, output from findDEG}
}
\value{
data frame with the spearman and pearson correlation estimates and
        p values for each cluster.
}
\description{
Creates a data frame with the spearman and pearson correlation pre/post
batch correction per cluster.
}
\examples{
data("deg_df")
corrDEG(deg_df)

}
