% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{deg_df}
\alias{deg_df}
\title{DEG dataframe, output of findDEG(ifnb_split, ifnb_split_corrected, clusterVars = c('seurat_clusters','seurat_clusters'), logFCThresh=2)}
\format{
dataframe with six columns
\describe{
\item{gene}{gene symbol}
\item{cluster}{cluster name}
\item{avg_log2FC_pre}{average log2FC of gene in cluster before correction}
\item{avg_log2FC_post}{average log2FC of gene in cluster after correction}
\item{delta}{avg_log2FC_pre - avg_log2FC_post}
\item{batch}{batch name}
}
}
\source{
IntegrationScore
}
\usage{
deg_df
}
\description{
DEG dataframe, output of findDEG(ifnb_split, ifnb_split_corrected, clusterVars = c('seurat_clusters','seurat_clusters'), logFCThresh=2)
}
\examples{
\dontrun{
 corrDEG
}
}
\keyword{datasets}
