% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDEG.R
\name{findDEG}
\alias{findDEG}
\title{Calculates the Pre/Post Batch Correction Differentially Expressed Genes}
\usage{
findDEG(srtObjPre, srtObjPost, batchNames, clusterVars, minPct, logFCThresh)
}
\arguments{
\item{srtObjPre}{Seurat object of the uncorrected data}

\item{srtObjPost}{Seurat object of the corrected data}

\item{batchNames}{character vector of length 2 with the batch names}

\item{clusterVars}{character vector of the cluster columns in Seurat objects,
these must be clusters obtained prior to batch correction
on each batch separately}

\item{minPct}{minimum percent for Seurat::FindMarkers}

\item{logFCThresh}{threshold for Seurat::FindMarkers}
}
\value{
data frame with DEG information for each batch pre and
        post correction
}
\description{
Creates a data frame with the DEGs of each batch pre and post correction.
}
