% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ifnb_corrected}
\alias{ifnb_corrected}
\title{corrected ifnb Seurat object list}
\format{
Seurat object
}
\source{
SeuratData
}
\usage{
ifnb_corrected
}
\description{
corrected ifnb Seurat object list
}
\examples{
\dontrun{
 findDEG
}
}
\keyword{datasets}
