% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ifnb_split}
\alias{ifnb_split}
\title{uncorrected ifnb Seurat object list}
\format{
Named List of 2 Seurat objects
}
\source{
SeuratData
}
\usage{
ifnb_split
}
\description{
uncorrected ifnb Seurat object list
}
\examples{
\dontrun{
 findDEG
}
}
\keyword{datasets}
