% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrationScore-package.R
\docType{package}
\name{IntegrationScore-package}
\alias{IntegrationScore}
\alias{IntegrationScore-package}
\title{IntegrationScore: Toolbox For Evaluating Batch Correction Methods}
\description{
The objective of this R package is to allow researchers working with scRNA-seq data to compare the effectiveness of batch correction methods.
}
\keyword{internal}
