% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDelta.R
\name{plotDelta}
\alias{plotDelta}
\title{Plot deltas by cluster of most altered genes under batch correction.}
\usage{
plotDelta(DEG_df, nTop)
}
\arguments{
\item{DEG_df}{data frame with DEG information for each batch pre and
post correction, output from findDEG}

\item{nTop}{integer specifying number of top genes to consider}
}
\value{
ggplot object
}
\description{
Plot deltas by cluster of most altered genes under batch correction.
}
