% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMAP.R
\name{plotUMAP}
\alias{plotUMAP}
\title{Plot UMAPs of pre and post correction.}
\usage{
plotUMAP(srtObjPre, srtObjPost, colorBy, nFeatures = 2000, nPC = 20)
}
\arguments{
\item{srtObjPre}{Seurat object(s) of the uncorrected data}

\item{srtObjPost}{Seurat object(s) of the corrected data}

\item{colorBy}{variable to color plot by}

\item{nFeatures}{number of top variable genes to use}

\item{nPC}{number of principle components to use}
}
\value{
ggplot object
}
\description{
Plot UMAPs of pre and post correction.
}
