% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIntegrationScore.R
\name{runIntegrationScore}
\alias{runIntegrationScore}
\title{Launch Shiny App for IntegrationScore}
\usage{
runIntegrationScore()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for IntegrationScore
The purpose of this app is
}
\examples{
\dontrun{

IntegrationScore::runIntegrationScore()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
