% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreDEG.R
\name{scoreDEG}
\alias{scoreDEG}
\title{Finds the correlation pre/post batch correction of the DEG by cluster.}
\usage{
scoreDEG(DEG_df, batchNames)
}
\arguments{
\item{DEG_df}{data frame with DEG information for each batch pre and
post correction, output from findDEG}

\item{batchNames}{character vector of length 2 with the batch names}
}
\value{
data frame with the spearman and pearson correlation estimates and
        p values for each cluster.
}
\description{
Creates a data frame with the spearman and pearson correlation pre/post
batch correction per cluster.
}
