% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreIntegration.R
\name{scoreIntegration}
\alias{scoreIntegration}
\title{Runs various test to evaluate batch correction}
\usage{
scoreIntegration(srtObjPre, srtObjPost)
}
\arguments{
\item{srtObjPre}{Seurat object of the uncorrected data}

\item{srtObjPost}{Seurat object of the corrected data}
}
\value{

}
\description{
Runs various test to evaluate batch correction
}
