% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDEG.R
\name{findDEG}
\alias{findDEG}
\title{Calculates the Pre/Post Batch Correction Differentially Expressed Genes}
\usage{
findDEG(
  srtObjPre,
  srtObjPost,
  batchVar,
  clusterVars = c("seurat_clusters", "seurat_clusters"),
  minPct = 0.15,
  logFCThresh = 1.5
)
}
\arguments{
\item{srtObjPre}{Named list of length 2 of 2 Seurat objects of the
uncorrected data split by batch, or single Seurat object to be split}

\item{srtObjPost}{Named list of length 2 of 2 Seurat objects of the corrected
data split by batch, or single Seurat object to be split}

\item{batchVar}{string of the batch variable column name}

\item{clusterVars}{character vector of the cluster columns in Seurat objects,
these must be clusters obtained prior to batch correction
on each batch separately}

\item{minPct}{minimum percent for Seurat::FindMarkers}

\item{logFCThresh}{threshold for Seurat::FindMarkers}
}
\value{
data frame with DEG information for each batch pre and
        post correction
}
\description{
Creates a data frame with the DEGs of each batch pre and post correction.
}
\examples{
library(Seurat)
data("ifnb_split")
data("ifnb_split_corrected")
deg_df <- findDEG(ifnb_split, ifnb_split_corrected, clusterVars = c('seurat_clusters','seurat_clusters'), logFCThresh=2)

}
